# dist/api.yml
openapi: 3.0.0
...
info:
  description: This is a simple API
  version: "1.0.0"
  title: Simple Inventory API
  # contact:
  #   email: you@your-company.com
  # license:
  #   name: Apache 2.0
  #   url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /auth/Register:
    post:
      tags:
        - User
      summary: A new user from the client is added
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: A new user has been added
        '409': 
          description: Wrong input parameters
  /auth/Login:
    post:
      tags:
        - User
      summary: A user from the client is signed in
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: A new user has been found
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        '401':
          description: Wrong input parameters
...
components:
  schemas:
    User:
      type: object
      description: User register
      properties:
        userID:
          type: integer
          description: user's id
          example: "3"
        username:
          type: string
          pattern: /[A-Za-z]/ #no numbers
          minLength: 3
          maxLength: 8
          description: user's username 
          example: "NoyaLev"
        firstName:
          type: string
          description: user's first name
          example: "Noya"
        lastName:
          type: string
          description: user's last name
          example: "Lev"
        country:
          type: string
          description: user's country
          example: "Israel"
        password:
          type: string
          description: password should contains between 5 to 10 characters, at least one of them is a number and a special char
          minLength: 5
          maxLength: 10
          format: password
          pattern:  /^(?=.*[0-9])(?=.*[a-zA-Z])([a-zA-Z0-9]+)$/
          example: "2y4$gd89"
        email:
          type: string
          format: email
          pattern:  /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/
          description: user's email
          example: "user123@gmail.com"
        image-url:
          type: string
          format: URL
          description: URL for user's image 
          example: "https://cloudinary.com/danilevy"
    UserLogin:
      type: object
      description: User login
      properties:
        username:
          type: string #no numbers
          # minLength: 3 
          # maxLength: 8 
          description: user's username 
          example: "NoyaLev"
        password: 
          type: string
          description: password
          format: password
          example: "df576!@dfg56"
...
  securitySchemes:
    cookieAuth:  
      type: apiKey
      in: cookie
      name: session  # cookie name


# openapi: 3.0.0
# info:
#   title: api
#   description: API for  project
#   version: 1.0.0
# tags:
#   - name: Users
#   - name: Auth
#   - name: League
#   - name: Players
#   - name: Teams
#   - name: Games
# paths:
#   /Register:
#     post:
#       tags:
#         - Auth
#       summary: Add new user
#       description: A new user from the client is added to the server
#       requestBody:
#         required: true
#         content:
#           application/json:
#             schema:
#               $ref: "#/components/schemas/User"
#       responses:
#         "201":
#           description: A new user has been added
#         "400":
#           description: wrong input parameters
#   /Login:
#     post:
#       tags:
#         - Auth
...